refactor(pubsub-open-telemetry): move publisher functions into a helper library
