+++ pubsub-open-telemetry/CMakeLists.txt
@@ -31,11 +31,15 @@ target_link_libraries(
     parse_args PUBLIC Boost::program_options google-cloud-cpp::pubsub
                       google-cloud-cpp::opentelemetry)
 
+add_library(publisher_helper STATIC publisher_helper.cc publisher_helper.h)
+target_compile_features(publisher_helper PUBLIC cxx_std_14)
+target_link_libraries(
+    publisher_helper PUBLIC parse_args)
+
 add_executable(publisher publisher.cc)
 target_compile_features(publisher PRIVATE cxx_std_14)
 target_link_libraries(
-    publisher PRIVATE google-cloud-cpp::pubsub google-cloud-cpp::opentelemetry
-                      parse_args)
+    publisher PRIVATE publisher_helper parse_args)
 
 add_executable(quickstart quickstart.cc)
 target_compile_features(quickstart PRIVATE cxx_std_14)
